+++ evmd/cmd/evmd/config/evm_app_options.go
@@ -4,11 +4,12 @@
 package config
 
 import (
-	evmconfig "github.com/cosmos/evm/config"
+	cosmosevmserverconfig "github.com/cosmos/evm/server/config"
 )
 
 // EvmAppOptions allows to setup the global configuration
 // for the Cosmos EVM chain.
 func EvmAppOptions(chainID uint64) error {
-	return evmconfig.EvmAppOptionsWithConfig(chainID, ChainsCoinInfo, cosmosEVMActivators)
+	evmCoinInfo := *cosmosevmserverconfig.DefaultEvmCoinInfo()
+	return EvmAppOptionsFromConfig(chainID, evmCoinInfo)
 }